+++ app/settings_form.py
@@ -1,4 +1,4 @@
-"""Application Settings Form.
+"""PytSite Blog Settings Form.
 """
 from pytsite import widget, lang, validation, settings
 
@@ -8,12 +8,17 @@ __license__ = 'MIT'
 
 
 class Form(settings.Form):
+    """PytSite Blog Settings Form.
+    """
+
     def _setup_widgets(self):
+        """Hook.
+        """
         self.add_widget(widget.input.StringList(
             uid='setting_links',
             weight=10,
-            label=lang.t('links'),
-            add_btn_label=lang.t('add_link'),
+            label=lang.t('app@links'),
+            add_btn_label=lang.t('app@add_link'),
             unique=True,
             rules=validation.rule.Url(),
         ))